Updates to add name

+++ app.py
@@ -67,6 +67,7 @@ def getTracks():
         auth=token_info['access_token'],
     )
 
+    current_user_name = sp.current_user()['display_name']
     short_term = sp.current_user_top_tracks(
         limit=10,
         offset=0,
@@ -83,9 +84,10 @@ def getTracks():
         time_range=LONG_TERM,
     )
 
-    os.remove(".cache")
+    if os.path.exists(".cache"): 
+        os.remove(".cache")
 
-    return render_template('receipt.html', short_term=short_term, medium_term=medium_term, long_term=long_term, currentTime=gmtime())
+    return render_template('receipt.html', user_display_name=current_user_name, short_term=short_term, medium_term=medium_term, long_term=long_term, currentTime=gmtime())
 
 
 @app.template_filter('strftime')
